Test Refactor - YamlServletProfileInitializerTest
- Use SystemPropertiesCleanupExtension to clean up system properties

[#169991138]
